+++ src/types/react-phone-number-input.d.ts
@@ -1,0 +1,4 @@
+declare module 'react-phone-number-input/locale/en' {
+  const enLocale: { [key: string]: string };
+  export default enLocale;
+}
